Agri Shop Application
Agri Shop Application is an app that helps farmers to get their Equipment through this app where
the app provides Seeds, Fertilizers, Pesticides, Plant Growth Regulators, Irrigation and Farming
Tools and Equipment that are required in agriculture. The admin Will Upload the Equipment
Photos with the details about it and it will also include the types of crops and also the farming
types like for large scale or for small scale. On the Other Farmer can view the product and can add
to the Cart, they can place their order even through a call. Through these app the farmers can easily
order the equipment whatever is required. The developed android application is an ecommerce app
which is used by both admin and farmers/Users.